+++ public/index.html
@@ -22,6 +22,6 @@
       <section class="footer__statistics"></section>
     </footer>
 
-    <script src="../src/main.js"></script>
+    <script src="./bundle.js"></script>
   </body>
 </html>
