+++ src/view/cards-container.js
@@ -1,0 +1,5 @@
+export const createCardsContainer = () => {
+  return (
+    `<div class="films-list__container"> </div>`
+  );
+};
