+++ src/view/main-section.js
@@ -1,0 +1,9 @@
+export const createMainSection = () => {
+  return (
+    `<section class="films">
+        <section class="films-list">
+            <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
+        </section>
+    </section>`
+  );
+};
