+++ src/view/most-commented-section.js
@@ -1,0 +1,7 @@
+export const createMostCommentedSection = () => {
+  return (
+    `<section class="films-list--extra">
+        <h2 class="films-list__title">Most commented</h2>
+    </section>`
+  );
+};
