+++ src/view/statistics.js
@@ -1,0 +1,5 @@
+export const createStatistics = () => {
+  return (
+    `<p>130 291 movies inside</p>`
+  );
+};
