+++ src/view/top-section.js
@@ -1,0 +1,7 @@
+export const createTopSection = () => {
+  return (
+    `<section class="films-list--extra">
+      <h2 class="films-list__title">Top rated</h2>
+    </section>`
+  );
+};
